passing tests

+++ test/modules/TestResultCoordinator/Reports/DirectMethod/Connectivity/DirectMethodConnectivityReport.cs
@@ -78,7 +78,7 @@ namespace TestResultCoordinator.Reports.DirectMethod.Connectivity
 
             // This tolerance is needed because sometimes we see large numbers of NetworkOnFailures.
             // When this product issue is resolved, we can remove this failure tolerance.
-            bool areNetworkOnFailuresWithinThreshold = (this.NetworkOnFailure / totalResults) < .30;
+            bool areNetworkOnFailuresWithinThreshold = ((double)this.NetworkOnFailure / totalResults) < .30;
 
             return this.MismatchFailure == 0 && this.NetworkOffFailure == 0 && areNetworkOnFailuresWithinThreshold && (this.NetworkOnSuccess + this.NetworkOffSuccess + this.NetworkOnToleratedSuccess + this.NetworkOffToleratedSuccess > 0);
         }
